Rewrote concurrent function execution handling
Using bash 5.1 or newer allows for getting a process id using the '-p' flag for
the wait built-in.  Using 'wait -p -n' allows for processing logs/error messages
on a first come first serve basis which is more efficient than a queue and allows
for functions to be retried if they fail fast.

Signed-off-by: Bob Schmitz III <14095796+rgschmitz1@users.noreply.github.com>